requirements.txt
fastapi[standard]
uvicorn
python-telegram-bot
google-generativeai
jinja2
python-dotenv
sqlmodel
